chore: add `resolved` label (#61416)
### What?

- Add a comment to issues labeled with `resolved`

Note to maintainers: it won't autoclose the issue. (This would need to
be added to [`nissuer`](https://github.com/balazsorban44/nissuer) as a
new feature)

Rendered comment
[here](https://github.com/vercel/next.js/blob/b868cf95f29a305bca1afabdb8db76e64a1cce30/.github/good-first-issue.md)

- I also reorganized the label comments to its own folder since we have
a few now.

### Why?

After verifying if an open issue has been fixed, a common task is to
comment on it and tell people to upgrade or open a new issue if it
persists. (Ideally, PRs would reference every issue they fix, but
sometimes we need to do this chore manually)

### How?

Using `nissuer`'s
[comment-label](https://github.com/balazsorban44/nissuer#label-management).


Closes NEXT-2306