Correctly tag stack frames as server frames
This implements a basic test for determining if a stack frame originated on the server.

Unfortunately, when RSC replays server errors on the client, it does not carry over the [0] symbol used to annotate server frames. Instead, errors are recreated by React and only include the message and stack.

This allows more tests to pass, as we are able to correctly trace stack frames by loading the appropriate server chunk.

https://github.com/vercel/next.js/blob/13af19a5627f2e2e538e9db9a6bfa0d13f216a72/packages/react-dev-overlay/src/internal/helpers/nodeStackFrames.ts#L33-L41