Fix closing a modal example in parallel routes docs (#61819)
The docs mention using a `Link` to close the modal, but navigating away
from a slot that was active won't automatically unmount unless it
matches with a new page for that slot. This is because client-side
navigations don't switch slots from active -> default. Default is only
used on a hard navigation.

I believe this example used to be in here but got removed, so I re-added
with some extra clarity.


Closes NEXT-2405