Updates app-index to only ever construct the initial data response once
There was a bug where if the root hydrates and then an update happens the intitial server data response can be replaced by one that is completely empty and will never resolve. This can lead to a frozen hydration that blocks interactivity. This udpate makes it so it is impossible for the initial data response to ever be created more than once.