Include Object.hasOwn polyfill
Seems like Next.js doesn't include the polyfill for Object.hasOwn out of the box, this has led to exception on a few (not fairly old ) broiwsers. 

Raising this smol PR in hope this would be covered in future versions. 

Here is the browser compatibility matrix for Object.hasOwn https://caniuse.com/mdn-javascript_builtins_object_hasown