Rename internal utility naming for clarification  (#62048)
`createActionProxy` isn't the accurate name here - the function doesn't
return anything new but only mark the value as registered. And also, it
might not be a proxy (which is implementation detail) and might not be
used for Server Actions only. `registerServerReference` is a better name
here.

There's no functionality change in this PR, only a few renames.

Closes NEXT-2490